Initial Commit
+AddCurrency - This class is simply a handler for quickly adding values
to the current value. This will add money to the instance it is
inheriting from, thus if we want to set the value for another instance
of Money, we will use the instance of addCurrency in Money
+Bill - A simple class in order to tag bills used in the transaction
(This is just a test class, unsure if it will be implemented)
+Coin - A simple class in order to tag coins used in the transaction
(This is just a test class, unsure if it will be implemented)
+CurrencyGenerator - This will randomly generate the bills and coins to
be used in the transaction (Not fully implemented yet)
+Game - This class holds all of the GameStates for the different games.
It also contains the currentGameState. This class can house other
properties that will be prevalent in each game
+Money - The main class that uses Bills and Coins (Though this simply
just uses test constructers as there is only one of each, unsure if
this will be used in the way I am thinking right now) in order to keep
track of the current money. The value of each currency is kept in this
class, we may put it in AddCurrency though
+MoneyIns - Short for Money Inspector, which simply is a test class to
make sure we are adding and returning what should be our value